chore: changes to meet php>=7.2.5 requirements
Since we are deprecating php<7.2.5 some code changes are done here, due to that a bump for some dependencies is also done in this commit. One example is phpunit which requires that test units extends from PHPUnit\Framework\TestCas and no from \PHPUnit_Framework_TestCase. Another example is that the annotations @expectedException, @expectedExceptionMessage, @expectedExceptionMessageRegExp, among others, are deprecated in phpunit 9, and instead the code had to be refactored to use the recommended function for each annotation, such as instead of @expectedException we need to use expectException(exceptionClassName), etc.